Add fallback nodes to connection
Sometimes the nodes would respond with 404s or 500 codes or are not
reachable due to other reasons. The connect function will now try
multiple nodes (at the moment two) before telling the user that there
was a problem. Additionally the function checks whether the networkID
matches the expected one to verify that the node really works.